+++ mainservice/src/main/java/com/example/mainservice/controllers/TeacherControllerREST.java
@@ -1,18 +1,27 @@
 package com.example.mainservice.controllers;
 
+import com.example.mainservice.exceptions.FailedToGetException;
 import com.example.mainservice.exceptions.teacher.TeacherAlreadyExistsException;
 import com.example.mainservice.exceptions.teacher.TeacherIllegalArgumentException;
 import com.example.mainservice.exceptions.teacher.TeacherNotFoundException;
 import com.example.mainservice.models.Teacher;
 import com.example.mainservice.services.interfaces.TeacherService;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.ResponseErrorHandler;
+import org.springframework.web.client.RestTemplate;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -21,7 +30,18 @@ import java.util.Map;
 @RequestMapping("/REST/teachers")
 public class TeacherControllerREST {
 
+    private static final String LESSON_SERVICE_URL = "https://rude-icons-flash-94-244-36-128.loca.lt/REST/";
+
     private TeacherService teacherService;
+
+    RestTemplate restTemplate = new RestTemplateBuilder().errorHandler(new ResponseErrorHandler() {
+        @Override
+        public boolean hasError(ClientHttpResponse clientHttpResponse) throws IOException {
+            return false;
+        }
+        @Override
+        public void handleError(ClientHttpResponse clientHttpResponse) throws IOException { }
+    }).build();
 
     @Autowired
     public TeacherControllerREST(TeacherService teacherService)
@@ -52,10 +72,55 @@ public class TeacherControllerREST {
 
     @DeleteMapping("/{id}")
     public Map<String, Boolean> deleteTeacher(@PathVariable(value = "id") Long id) throws TeacherNotFoundException {
+        String url = LESSON_SERVICE_URL + "lessons/deleteByTeacherId/" + id.toString();
+        ResponseEntity<String> response = restTemplate.postForEntity(url, "", String.class);
+        if(response.getStatusCode() != HttpStatus.OK)
+            throw new FailedToGetException("Failed to delete lessons of teacher with id " + id.toString() + "!");
         teacherService.deleteTeacher(id);
         Map<String, Boolean> result = new HashMap<>();
         result.put("deleted",true);
         return result;
+    }
+
+    @GetMapping("/withLessons/{id}")
+    public Map<String, Object> getTeacherAndAllHisLessonsById(@PathVariable(value = "id") Long id) throws Exception {
+        Teacher teacher = teacherService.getTeacherById(id);
+        ObjectMapper mapper = new ObjectMapper();
+        TypeReference<HashMap<String,Object>> typeRef = new TypeReference<>() {};
+        Map<String, Object> teacherMap = mapper.convertValue(teacher,typeRef);
+        List<HashMap<String,Object>> lessonsMaps = getAllLessonsByTeacherId(teacher.getId());
+        HashMap<String,Object> map = new HashMap<>();
+        map.put("teacher",teacherMap);
+        map.put("lessons",lessonsMaps);
+        return map;
+    }
+
+    private List<HashMap<String,Object>> getAllLessonsByTeacherId(Long id) throws JsonProcessingException {
+        String url = LESSON_SERVICE_URL + "lessons/teacherId/" + id.toString();
+        ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
+        if(response.getStatusCode() != HttpStatus.OK)
+            throw new FailedToGetException("Failed to fetch lessons of teacher with id " + id.toString() + "!");
+        ObjectMapper mapper = new ObjectMapper();
+        TypeReference<List<HashMap<String,Object>>> typeRef = new TypeReference<>() {};
+        return mapper.readValue(response.getBody(),typeRef);
+    }
+
+    @PostMapping("/deleteLessonFromTeacher/{teacherId}/{lessonId}")
+    public Teacher deleteTeacherLesson(@PathVariable(value = "teacherId") Long teacherId, @PathVariable(value = "lessonId") Long lessonId) {
+        Teacher teacher = teacherService.getTeacherById(teacherId);
+        List<Long> lessons = teacher.getLessons();
+        lessons.remove(lessonId);
+        teacher.setLessons(lessons);
+        return teacherService.updateTeacher(teacher);
+    }
+
+    @PostMapping("/addLessonForTeacher/{teacherId}/{lessonId}")
+    public Teacher addTeacherLesson(@PathVariable(value = "teacherId") Long teacherId, @PathVariable(value = "lessonId") Long lessonId) {
+        Teacher teacher = teacherService.getTeacherById(teacherId);
+        List<Long> lessons = teacher.getLessons();
+        lessons.add(lessonId);
+        teacher.setLessons(lessons);
+        return teacherService.updateTeacher(teacher);
     }
 
     @ExceptionHandler(TeacherNotFoundException.class)
